requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Telegram Bot
python-telegram-bot>=20.7

# Database
# sqlite3 is built into Python

# Utilities
python-dateutil>=2.8.2

# Development/Testing
pytest>=7.4.0
pytest-mock>=3.12.0
